add float for mass and greate Pk in single precision

+++ cosmology/GravityParticleContainer.hpp
@@ -19,7 +19,7 @@ public:
     /**
      * @brief Mass of the particle.
      */
-    ippl::ParticleAttrib<double> m;
+  ippl::ParticleAttrib<float> m;
 
     /**
      * @brief Velocity of the particle.
